fix(estimator): seven places where a name or a document disagreed with the code
``_refit_noise`` said "the series is innovations" 50 lines under
``_note_residual``'s "NOT the innovation" and its account of why. One file
carried two opposite descriptions of one series; the stale one is corrected.

Spec section 7 promised a per-tier last residual and last sigma in the
published witnesses. Neither reached the solution, and
``Admitter.last_residual_ns`` was written, cleared, and read by nothing.
Both now appear in ``TimingSolution.witnesses``, whose value type widens to
float and whose deep freeze is unchanged. A tier that has only ever spoken
about RATE carries neither key: its innovation is nanoseconds per second
and its sigma parts per million, and absent beats mislabelled.

``q1`` does no work in the assembled system -- ``noise_from_standin``
returns zero and ``_refit_noise`` discards the fitted value -- and the fit
site now says so, and why: the short tau is where witness noise dominates,
and a q1 read there credited a governed ruler with 866 microseconds of
phase wander a minute.

``_DIMENSIONLESS_TO_NS2`` was unused. Gone.

``TimingSolution.utc_ns_at`` rounded the phase and the projection
separately where spec section 1 rounds the sum once, costing up to a
nanosecond. It rounds once now, as ``ClockState.utc_ns_at`` always has.

``q_source`` was sticky. A ``RulerNoise``'s ``source`` string travels with
the coefficient and outlives the fit that set it, so the published label
described a past measurement rather than the number the filter runs on. It
is derived from the coefficient now: q2 at the declared floor reads
``standin``, whatever an earlier fit called itself. A provenance field that
can disagree with the value it describes undoes the honesty argument the
rest of this design rests on.

And the comment at ``rate_not_positive`` still claimed ruling R28's
exemption covered only ``not_finite``, so an impossible state would "raise
loudly". R37 widened it and the solution withholds instead.

The four spurious ``# noqa: E501`` suppressions in ``process_noise.py``
sat on lines that were never too long. Gone.

+++ src/hf_timestd/estimator/solution.py
@@ -50,7 +50,11 @@ class TimingSolution:
     covariance: tuple[float, float, float]
     verdict: str
     refusal: str | None
-    witnesses: Mapping[str, Mapping[str, int]]
+    # Per tier: ``accepted`` and ``rejected`` counts, and -- for a tier that
+    # has supplied at least one phase observation -- ``last_residual_ns``
+    # and ``last_sigma_ns``. Floats, because the last two are, and a mapping
+    # whose value type depended on the key would be worse (spec section 7).
+    witnesses: Mapping[str, Mapping[str, float]]
     a_level: str
     ruler_provenance: str
     q_source: str
